python-backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.3
contourpy==1.3.2
cramjam==2.11.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
espn-api==0.45.1
exceptiongroup==1.3.0
executing==2.2.0
fastparquet==2024.11.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.59.1
fsspec==2025.7.0
gotrue==2.12.3
greenlet==3.2.4
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.2
langsmith==0.4.14
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nfl_data_py==0.3.3
numpy==1.26.4
openai==1.99.9
orjson==3.11.2
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==25.0
pandas==1.5.3
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
playwright==1.54.0
postgrest==1.1.1
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
realtime==2.7.0
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.0
supafunc==0.10.1
tavily-python==0.7.10
tenacity==9.1.2
tiktoken==0.11.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
zstandard==0.23.0
